magic one
for some reason, i need to use English this time. the arm is about magic, it's belongs to near fight ,and the attack time will last for 5 seconds, then the magic square will disapare. there are some bugs, 1. the hero is below the magic square, for the layer of the hero seems too low, but i can't fine the define of it 2. it is fadeout but not fadein
i wish it is okay for you to use i push, i have changed it into win32-build, and .gitignore has been changed